No longer auto-removes the Armature modifier
While it hasn't come up so far, this step carried a risk of a user breaking their rig without realizing it, if they have already linked it to the Tower Unite armature. 

The reason it was added in the first place was to save a little bit of effort, but it's just an overreach of the script's purpose. Doing this properly is now covered by a video guide linked in the README file.