Remove "Delete rig" hint
The hint to delete the rig, before selecting the mesh, makes it seem like a required step, which it isn't. The video guide also covers using the rig to pose the hands before removing it and replacing it with the Tower Unite armature, making it a confusing step. (A user might wonder "When do you do it?")

The reason it was added was because this script was made for my friends, and this step served as a useful reminder to them to just get rid of the rig as early as possible. Now that I rely on the video guide, which encourages using it for posing the hands first, it's potentially confusing.